+++ src/main/haskell/kore/src/Kore/AST/MetaOrObject.hs
@@ -33,6 +33,8 @@ import Data.Proxy
 import GHC.Generics
        ( Generic )
 
+import Kore.Unparser
+
 toProxy :: a -> Proxy a
 toProxy _ = Proxy
 
@@ -89,6 +91,15 @@ deriving instance (OrdMetaOrObject thing) => Ord (Unified thing)
 deriving instance (ShowMetaOrObject thing) => Show (Unified thing)
 
 instance (NFData (thing Meta), NFData (thing Object)) => NFData (Unified thing)
+
+instance
+    (forall level. Unparse (thing level)) =>
+    Unparse (Unified thing)
+  where
+    unparse =
+        \case
+            UnifiedMeta meta -> unparse meta
+            UnifiedObject object -> unparse object
 
 {-|Given a function transforming objects of 'Meta' type and another transforming
 objects of 'Object' type, 'applyUnified' builds the corresponding direct sum
